test_task_2_1/requirements.txt
pytest==8.3.4
requests==2.32.3
jsonschema==4.17.3
allure-pytest==2.8.19
